Assumed interactions
Looking at the HC set, we assume every domain on each interacting gene
interacts with each domain on the other interacting gene.  This new set
presumably contains new DDIs that are not specified in the iPfam set,
since the iPfam set only accounted for ~20% of the interactions in the
HC set.