All DDIs in the HC set
An exhaustive list of all domain-domain combinations between source and
target in the HC set.  There is no evidence that these are true DDIs,
however some of them likely are true novel DDIs based on the fact that
iPfam cannot predict the majority of known gene-gene interactions in the
HC set.