All DDIs in the HC set, now with counts
An exhaustive list of all domain-domain combinations between source and
target in the HC set.  There is no evidence that these are true DDIs,
however some of them likely are true novel DDIs based on the fact that
iPfam cannot predict the majority of known gene-gene interactions in the
HC set.  This file also states how many of the gene-gene interactions in
the HC set have this DDI.